Muevo funciones de fechas a módulo aparte
