+++ ckanext/gobar_theme/helpers/__init__.py
@@ -6,3 +6,4 @@ from .licenses import *
 from .types import *
 from .cron import *
 from .distribution import *
+from .dates import *
