+++ ckanext/gobar_theme/helpers/dates.py
@@ -1,0 +1,40 @@
+#!coding=utf-8
+from datetime import time
+import moment
+from dateutil import parser, tz
+import ckan.lib.helpers as ckan_helpers
+from ckan.common import _
+
+
+
+def convert_iso_string_to_utc(date_string=''):
+    if date_string is None:
+        return ''
+    try:
+        date_time = parser.parse(date_string)
+    except ValueError:
+        # date_string es un string inválido o None
+        return ''
+    if date_time.time() == time(0):
+        return date_string
+    if date_time.tzinfo is not None:
+        utc_date_time = date_time.astimezone(tz.tzutc())
+    else:
+        utc_date_time = date_time
+    utc_date_time = utc_date_time.replace(tzinfo=None)
+    return utc_date_time.isoformat()
+
+
+def date_format_to_iso(date):
+    if date:
+        return moment.date(date, "%d/%m/%Y").isoformat()
+    return date
+
+
+def render_ar_datetime(datetime_):
+    datetime_ = ckan_helpers._datestamp_to_datetime(convert_iso_string_to_utc(datetime_))
+    if not datetime_:
+        return ''
+    return _('{day} de {month} de {year}').format(year=datetime_.year,
+                                                  month=datetime_.month,
+                                                  day=datetime_.day)
